requirements.txt
openai==1.98.0
google-generativeai==0.8.5
jsonlines